# internal/features/user_settings/dto/user_settings_dto.go
package dto

type SettingItem struct {
	Key   string `json:"key" validate:"required"`
	Value string `json:"value" validate:"required"`
}

type UpsertSettingsRequest struct {
	Settings []SettingItem `json:"settings" validate:"required,dive"`
}

type SettingResponse struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}